+++ tests/ct/ct_che_make_moves.h
@@ -9,7 +9,7 @@ void che_make_moves_tested_with_long_game_1();
 void che_make_moves_tested_with_long_game_2();
 void che_make_moves_tested_with_long_game_3();
 void ct_perft_v1( const char *root, int depth, int expected_nc, bool progress );
-void ct_perft_v2( const Pos *root, int depth, int expected_nc, bool progress );
+void ct_perft_v2( const Pos *root, int depth, long long expected_nc, bool progress );
 
 #endif
 // end CT_CHE_MAKE_MOVES_H
